move ConsoleWriterFactory implementation in util.cpp
