+++ src/util.cpp
@@ -1209,6 +1209,31 @@ QString CCountyFlagIcons::GetResourceReference ( const QLocale::Country eCountry
 }
 
 
+// Console writer factory ------------------------------------------------------
+QTextStream* ConsoleWriterFactory::get()
+{
+    if ( ptsConsole == nullptr )
+    {
+#if _WIN32
+        if ( !AttachConsole ( ATTACH_PARENT_PROCESS ) )
+        {
+            // Not run from console, dump logging to nowhere
+            static QString conout;
+            ptsConsole = new QTextStream ( &conout );
+        }
+        else
+        {
+            freopen ( "CONOUT$", "w", stdout );
+            ptsConsole = new QTextStream ( stdout );
+        }
+#else
+        ptsConsole = new QTextStream ( stdout );
+#endif
+    }
+    return ptsConsole;
+}
+
+
 /******************************************************************************\
 * Global Functions Implementation                                              *
 \******************************************************************************/
